Add OOP for parser
